Support the new GA topology.kubernetes.io/zone label
With Kubernetes 1.17+ the 'failure-domain.beta.kubernetes.io/zone' label is deprecated
in favor of the new 'topology.kubernetes.io/zone' label, so prefer that here as well.

Especially when migrating clusters there's a good chance both labels are used for a while
in parallel, so we cannot just remove the old label support here and require a flag-day
node upgrade.